fix(#180): 맵 관리 일괄 저장 트랜잭션 범위 개선
