+++ src/main/java/io/github/ascrew/monomatbe/domain/map/dto/ManageMapRequest.java
@@ -23,10 +23,21 @@ public record ManageMapRequest(
 
         boolean isPublic,
 
+        /**
+         * 저장 후 활성 상태로 남아야 하는 전체 문제 목록
+         * 변경된 문제만 보내는 필드가 아니다.
+         *
+         * <p>기존 활성 문제는 이 목록에 포함되거나 {@code deletedItemIds}에 포함되어야 한다.
+         * 신규 문제는 {@code id = null}로 전달한다.</p>
+         */
         @NotNull(message = "문제 목록은 필수입니다.")
         @Size(max = MapItemPolicy.MAX_ITEMS_PER_MAP, message = "한 맵에 등록할 수 있는 문제 수를 초과했습니다.")
         List<@Valid ManageMapItemRequest> items,
 
+        /**
+         * 기존 활성 문제 중 삭제할 문제 ID 목록입니다.
+         * {@code items[].id}와 중복될 수 없습니다.
+         */
         List<@Positive(message = "삭제할 문제 ID는 양수여야 합니다.") Long> deletedItemIds
 ) {
 }
